backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# AI & ML
openai==1.6.1
langchain==0.0.350
langchain-openai==0.0.2
pinecone
# pinecone-client==2.2.4

# Audio Processing
pydub==0.25.1
python-multipart==0.0.6
gtts==2.4.0

# Utilities
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
httpx==0.25.2
reportlab==4.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
